게시글 인증 인가처리

+++ src/main/java/org/example/demo_ssr_v1_1/board/BoardController.java
@@ -1,14 +1,13 @@
 package org.example.demo_ssr_v1_1.board;
 
-import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
+import org.example.demo_ssr_v1_1.user.User;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 
 import java.util.List;
 
@@ -19,24 +18,62 @@ public class BoardController {
 
     private final BoardPersistRepository repository;
 
-    // 게시글 수정 폼 페이지 요청
-    // http://localhost:8080/board/1/update
+    /**
+     * 게시글 수정 화면 요청
+     * @param id
+     * @param model
+     * @param session
+     * @return
+     */
     @GetMapping("/board/{id}/update")
-    public String updateForm(@PathVariable Long id,Model model) {
+    public String updateForm(@PathVariable Long id,Model model, HttpSession session) {
 
+       // 1. 인증 검사 (0)
+       User sessionUser = (User)session.getAttribute("sessionUser"); // sessionUser -> 상수
+       if(sessionUser == null) {
+           System.out.println("로그인 안한 사용자의 요청이 들어 옴");
+           return "redirect:/login";
+       } 
+
+       // 2. 인가 검사 (0)
        Board board =  repository.findById(id);
        if(board == null) {
-           throw new RuntimeException("수정할 게시글을 찾을 수 없어요");
+            throw new RuntimeException("수정할 게시글을 찾을 수 없어요");
        }
+
+        if(board.isOwner(sessionUser.getId()) == false) {
+            System.out.println("님이 작성한 게시글이 아니에요");
+            return "redirect:/";
+        }
+
        model.addAttribute("board", board);
        return "board/update-form";
     }
 
-    // 게시글 수정 요청 (기능요청)
-    // http://localhost:8080/board/1/update
+    /**
+     * 게시글 수정 요청 기능
+     * @param id
+     * @param updateDTO
+     * @param session
+     * @return
+     */
     @PostMapping("/board/{id}/update")
     public String updateProc(@PathVariable Long id,
-                             BoardRequest.UpdateDTO updateDTO) {
+                             BoardRequest.UpdateDTO updateDTO, HttpSession session) {
+
+        // 1. 인증 처리 (o)
+        User sessionUser =  (User)session.getAttribute("sessionUser");
+        if(sessionUser == null) {
+            System.out.println("로그인 안한 사용자가 요청 함");
+            return "redirect:/login";
+        }
+
+        Board board = repository.findById(id);
+        if(board.isOwner(sessionUser.getId()) == false) {
+            System.out.println("님이 작성한 게시글이 아니에요");
+            return "redirect:/";
+        }
+
         try {
             repository.updateById(id, updateDTO);
             // 더티 체킹 활용
@@ -45,8 +82,13 @@ public class BoardController {
         }
         return "redirect:/board/list";
     }
-    
-    // 게시글 목록 요청
+
+
+    /**
+     * 게시글 목록 화면 요청
+     * @param model
+     * @return
+     */
     @GetMapping({"/board/list", "/"})
     public String boardList(Model model) {
         List<Board> boardList = repository.findAll();
@@ -54,45 +96,79 @@ public class BoardController {
         return "board/list";
     }
 
-    // 게시글 저장 화면 요청
-    // http://localhost:8080/board/save
+
+    /**
+     * 게시글 작성 화면 요청
+     * @param session
+     * @return
+     */
     @GetMapping("/board/save")
-    public String saveFrom() {
+    public String saveFrom(HttpSession session) {
+        User sessionUser = (User) session.getAttribute("sessionUser");
+        if(sessionUser == null) {
+            return "redirect:/login";
+        }
         return "board/save-form";
     }
 
-    // 게시글 저장 요청 (기능 요청)
-    // http://localhost:8080/board/save
-    //
+    /**
+     * 게시글 작성 요청 기능
+     * @param saveDTO
+     * @param session
+     * @return
+     */
     @PostMapping("/board/save")
-    public String saveProc(BoardRequest.SaveDTO saveDTO) {
-        // HTTP 요청 : username=값&title=값&content=값
-        // 스프링이 처리 : new SaveDTO(), setter 메서드 호출해서 값을 쏙 ~ 넣어줌
-        Board board = saveDTO.toEntity();
+    public String saveProc(BoardRequest.SaveDTO saveDTO, HttpSession session) {
+        // 1. 인증 처리 확인
+        User sessionUser = (User) session.getAttribute("sessionUser");
+        if(sessionUser == null) {
+            return "redirect:/login";
+        }
+
+        Board board = saveDTO.toEntity(sessionUser);
         repository.save(board);
         return "redirect:/";
     }
 
-    // 삭제 @DeleteMapping 이지만 form 태그 활용 없음 get, post (fetch 함수 활용)
+    /**
+     * 게시글 삭제 요청 기능
+     * @param id
+     * @param session
+     * @return
+     */
     @PostMapping("/board/{id}/delete")
-    public String delete(@PathVariable Long id) {
+    public String delete(@PathVariable Long id, HttpSession session) {
+        // 1. 인증 처리 (o)
+        // 1. 인증 처리 확인
+        User sessionUser = (User) session.getAttribute("sessionUser");
+        if(sessionUser == null) {
+            return "redirect:/login";
+        }
+        // 2. 인가 처리 (o) || 관리자 권한
+        Board board = repository.findById(id);
+        if(board.isOwner(sessionUser.getId()) == false) {
+            System.out.println("님이 작성한 게시글이 아니에요");
+            return "redirect:/";
+        }
+
         repository.deleteById(id);
         return "redirect:/";
     }
 
-    // 상세보기 화면
-    // http://localhost:8080/board/1
+    /**
+     * 게시글 상세 보기 화면 요청
+     * @param id
+     * @param model
+     * @return
+     */
     @GetMapping("board/{id}")
     public String detail(@PathVariable Long id, Model model) {
-
         Board board = repository.findById(id);
         if(board == null) {
             // 404
             throw new RuntimeException("게시글을 찾을 수 없어요 : " + id);
         }
-
         model.addAttribute("board", board);
-
         return "board/detail";
     }
